requirements.txt
crewai
langchain_groq
streamlit
python-dotenv
langchain_openai
crewai[tools]
fpdf
